% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B0}
\alias{B0}
\title{B0 is a 10 x 10 matrix representing the upper triangle of a actuarial
development triangle used by Roger Hayne to illustrate the stochastic
loss reserving software in}
\format{
A 10 x 10 numeric matrix
\describe{
\item{B0}{Development triangle.}
}
}
\source{
"A Flexible Framework for Stochastic Reserving Models" found at
\url{https://variancejournal.org/article/120823}
}
\usage{
B0
}
\description{
B0 is a 10 x 10 matrix representing the upper triangle of a actuarial
development triangle used by Roger Hayne to illustrate the stochastic
loss reserving software in
}
\keyword{datasets}
