% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berquist.R
\name{berquist}
\alias{berquist}
\title{Create list for Berquist-Sherman incremental severity model}
\usage{
berquist(tau, B0, ptd, msk)
}
\arguments{
\item{tau}{do not know}

\item{B0}{development triangle}

\item{ptd}{do not know}

\item{msk}{mask for triangle}
}
\description{
g - Assumed loss emergence model, a function of the parameters a.
Note g must be matrix-valued with 10 rows and 10 columns
}
\details{
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is the Berquist-Sherman incremental severity model
}
