% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capecod.R
\name{capecod}
\alias{capecod}
\title{Create list for Kramer Chain Ladder parmaterization model
g - Assumed loss emergence model, a function of the parameters a.
Note g must be matrix-valued with 10 rows and 10 columns
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is nonlinear and based on the Kramer Chain Ladder parmaterization}
\usage{
capecod(tau, B0, ptd, msk)
}
\arguments{
\item{tau}{do not know}

\item{B0}{development triangle}

\item{ptd}{do not know}

\item{msk}{mask for triangle}
}
\description{
Create list for Kramer Chain Ladder parmaterization model
g - Assumed loss emergence model, a function of the parameters a.
Note g must be matrix-valued with 10 rows and 10 columns
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is nonlinear and based on the Kramer Chain Ladder parmaterization
}
