% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Create list for Cape Cod model}
\usage{
chain(tau, B0, ptd, msk)
}
\arguments{
\item{tau}{do not know}

\item{B0}{development triangle}

\item{ptd}{do not know}

\item{msk}{mask for triangle}

\item{theta}{do not know}
}
\description{
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is a version of the Cape Cod model but with the restriction
that the expected cumulative averge payments to date equal the actual
average paid to date.  Because of this restriction the incremental averages
are expressed as a percentage times the expected ultimate by row.
Formulae all assume a full, square development triangle.
}
