% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_incremental_avg_matrix.R
\name{get_incremental_avg_matrix}
\alias{get_incremental_avg_matrix}
\title{Calculate incremental average matrix}
\usage{
get_incremental_avg_matrix(dTriangle)
}
\arguments{
\item{dTriangle}{development triangle}
}
\value{
incremental average matrix from development triangle
}
\description{
Calculate incremental average matrix
}
