% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoerl.R
\name{hoerl}
\alias{hoerl}
\title{Create list for Generalized Hoerl Curve Model with trend
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is Wright's operational time model with trend added}
\usage{
hoerl(tau, B0, ptd, msk)
}
\arguments{
\item{tau}{do not know}

\item{B0}{development triangle}

\item{ptd}{do not know}

\item{msk}{mask for triangle}
}
\description{
Create list for Generalized Hoerl Curve Model with trend
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g.obj must be flexible enough to handle
both.
Here g.obj is Wright's operational time model with trend added
}
