% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gradient_of_objective.R
\name{make_gradient_of_objective}
\alias{make_gradient_of_objective}
\title{Make Gradient of the objective function}
\usage{
make_gradient_of_objective(a, A, dnom, g_obj, g_grad)
}
\arguments{
\item{a}{do not know}

\item{A}{do not know}

\item{dnom}{numeric vector representing the exposures (claims) used in the
denominator}

\item{g_obj}{objective function}

\item{g_grad}{gradient function}
}
\description{
Make Gradient of the objective function
}
