% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_log_hessian.R
\name{make_log_hessian}
\alias{make_log_hessian}
\title{Make Hessian of the objective function}
\usage{
make_log_hessian(a, A, dnom, g_obj, g_grad, g_hess)
}
\arguments{
\item{a}{do not know}

\item{A}{do not know}

\item{dnom}{numeric vector representing the exposures (claims) used in the
denominator}

\item{g_obj}{objective function}

\item{g_grad}{gradient function}

\item{g_hess}{hessian function}
}
\description{
-   m_expected_value is the expectated value matrix
-   m_variance is the matrix of variances
-   A, m_expected_value, m_variance all have shape c(size, size)
-   The variables _v are copies of the originals to shape
    c(npar,size,size), paralleling the gradient of g.
-   The variables _m are copies of the originals to shape
    c(npar,npar,size,size), paralleling the hessian of g
hessian-of-objective-function
}
