% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_negative_log_likelihood.R
\name{make_negative_log_likelihood}
\alias{make_negative_log_likelihood}
\title{Make Negative Loglikelihood Function to be Minimized}
\usage{
make_negative_log_likelihood(a, A, dnom, g_obj)
}
\arguments{
\item{a}{do not know}

\item{A}{do not know}

\item{dnom}{numeric vector representing the exposures (claims) used in the
denominator}

\item{g_obj}{objective function}
}
\description{
Note that the general
form of the model has parameters in addition to those in the loss model,
namely the power for the variance and the constant of proprtionality that
varies by column.  So if the original model has k parameters with size
columns of data, the total objective function has k + size + 1 parameters
}
