% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_description.R
\name{model_description}
\alias{model_description}
\title{Get the long description of a model}
\usage{
model_description(model)
}
\arguments{
\item{model}{character vector of length one having the short name of the
model.}
}
\value{
Long description of a model as a character vector of length one.
}
\description{
Get the long description of a model
}
