% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAA_cumulative}
\alias{RAA_cumulative}
\title{RAA Cumulative Claims Triangle}
\format{
A 10 x 10 numeric matrix with:
\describe{
  \item{rows}{Accident years 1981-1990}
  \item{columns}{Development periods 1-10}
  \item{values}{Cumulative paid claims (thousands of dollars)}
}
}
\source{
Originally from the Reinsurance Association of America.
  Obtained from the ChainLadder R package (Gesmann et al., 2025).
  \url{https://mages.github.io/ChainLadder/}
}
\usage{
RAA_cumulative
}
\description{
Cumulative claims data from the Reinsurance Association of America (RAA),
obtained from the ChainLadder package. This is a classic 10x10 development
triangle commonly used in actuarial reserving literature.
}
\details{
The data represents historical general liability claims from 1981-1990 with
10 development periods. Values are in thousands of dollars.
}
\examples{
data(RAA_cumulative)
# View the triangle
print(RAA_cumulative)

# Convert to incremental
RAA_incr <- cbind(RAA_cumulative[,1],
                  t(apply(RAA_cumulative, 1, diff)))
}
\references{
Gesmann M, Murphy D, Zhang Y, Carrato A, Wuthrich M, Concina F, Dal Moro E
(2025). ChainLadder: Statistical Methods and Models for Claims Reserving
in General Insurance. R package version 0.2.20.
}
\seealso{
\code{\link{RAA_incremental}} for the incremental version
}
\keyword{datasets}
