% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAA_incremental}
\alias{RAA_incremental}
\title{RAA Incremental Claims Triangle}
\format{
A 10 x 10 numeric matrix with:
\describe{
  \item{rows}{Accident years 1981-1990}
  \item{columns}{Development periods 1-10}
  \item{values}{Incremental paid claims (thousands of dollars)}
}
}
\source{
Originally from the Reinsurance Association of America.
  Obtained from the ChainLadder R package (Gesmann et al., 2025).
  \url{https://mages.github.io/ChainLadder/}
}
\usage{
RAA_incremental
}
\description{
Incremental claims data from the Reinsurance Association of America (RAA),
converted from the cumulative triangle in the ChainLadder package. This is
a classic 10x10 development triangle commonly used in actuarial reserving
literature.
}
\details{
The data represents historical general liability claims from 1981-1990 with
10 development periods. Values are in thousands of dollars.
}
\note{
This data was converted from cumulative to incremental format.
  Negative values (e.g., in 1982, period 7) represent payment recoveries
  or adjustments.
}
\examples{
data(RAA_incremental)
# View the triangle
print(RAA_incremental)

# Convert back to cumulative
RAA_cum <- t(apply(RAA_incremental, 1, cumsum))
}
\references{
Gesmann M, Murphy D, Zhang Y, Carrato A, Wuthrich M, Concina F, Dal Moro E
(2025). ChainLadder: Statistical Methods and Models for Claims Reserving
in General Insurance. R package version 0.2.20.
}
\seealso{
\code{\link{RAA_cumulative}} for the cumulative version
}
\keyword{datasets}
