% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berquist.R
\name{berquist}
\alias{berquist}
\title{Create list for Berquist-Sherman incremental severity model}
\usage{
berquist(B0, paid_to_date, upper_triangle_mask)
}
\arguments{
\item{B0}{development triangle}

\item{paid_to_date}{numeric vector of length \code{size}. It is the lower
diagnal of
the development triangle in row order. It represents the amount paid to date.}

\item{upper_triangle_mask}{is a mask matrix of allowable data, upper
triangular assuming same development increments as exposure increments}
}
\description{
g - Assumed loss emergence model, a function of the parameters a.
Note g must be matrix-valued with size rows and size columns
}
\details{
g itself
Basic design is for g to be a function of a single parameter vector, however
in the simulations it is necessary to work on a matrix of parameters, one
row for each simulated parameter, so g_obj must be flexible enough to handle
both.
Here g_obj is the Berquist-Sherman incremental severity model
}
