% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnom}
\alias{dnom}
\title{dnom is count forcast for the B0 development triangle; the exposures
(claims) used in the denominator
#' @format integer vector of length 10
\describe{
\item{B0}{count forcast for the B0 development triangle.}
}}
\format{
An object of class \code{numeric} of length 10.
}
\source{
"A Flexible Framework for Stochastic Reserving Models" found at
\url{https://variancejournal.org/article/120823}
}
\usage{
dnom
}
\description{
dnom is count forcast for the B0 development triangle; the exposures
(claims) used in the denominator
#' @format integer vector of length 10
\describe{
\item{B0}{count forcast for the B0 development triangle.}
}
}
\keyword{datasets}
